refactor(types): break the AST/Utils include cycle
- Utils.hpp needed the full AST header only to name the builtin type
  identifiers, and it named them through the class, as
  BuiltinType::TypeId -- which is what dragged the whole header in.
  Types.hpp now holds the identifiers on their own, so Utils.hpp
  includes that instead, and varTypeToTypeId stays in Utils.cpp where
  the full AST is.
- The identifiers become a scoped enum on the way out. An unscoped one
  converts to int on its own, so a BuiltinTypeId could be compared
  against a width or an opcode and nothing would say so.

+++ src/AbstractSyntaxTree.cpp
@@ -1140,7 +1140,8 @@ llvm::Value* Constant::genCode(CodeGenerator& generator) {
     case BuiltinTypeId::BOOL:
       return g_builder.getInt1(boolValue_);
     default:
-      throw std::logic_error("const type " + std::to_string(typeId_) +
+      throw std::logic_error("const type " +
+                             std::to_string(static_cast<int>(typeId_)) +
                              " is not supported!");
   }
 }
